Derive patch, remove errors
the relevant part at this point is the patch that we derive.  Now I can
work with this to derive a patch file from the two plainTextEdit box's

I think I know what I need to do
well, not 100%, but I don't even want to run that line atm
and it apparently derives a patch
so... I can get me a patch!